% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form}
\alias{form}
\title{Create an HTML Form}
\usage{
form(action, method = "get", ...)
}
\arguments{
\item{action}{Where to forward the form information.}

\item{method}{One of \code{'get'} or \code{'post'}.}

\item{\ldots}{widget or input objects, see \code{\link{inputs}} and
\code{\link{widgets}}.}
}
\description{
An HTML form.
}

