% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{widget}
\alias{widget}
\alias{is.widget}
\title{HTML Widgets}
\usage{
widget(attrs = NULL)

is.widget(x)
}
\arguments{
\item{attrs}{Widget attributes.}
}
\description{
HTML widgets are most commonly used under-the-hood of the common HTML form
\code{\link{inputs}}. Custom HTML widgets can be created.
}
