% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R, R/textarea.R
\name{inputs}
\alias{inputs}
\alias{text_input}
\alias{email_input}
\alias{password_input}
\alias{number_input}
\alias{tel_input}
\alias{url_input}
\alias{hidden_input}
\alias{text_area}
\title{HTML Form Inputs}
\usage{
text_input(...)

email_input()

password_input()

number_input()

tel_input()

url_input()

hidden_input()

text_area()
}
\description{
Common HTML form inputs.
}
